requirements.txt
streamlit
opencv-python
pillow
numpy
requests
matplotlib
streamlit-webrtc
av
